+++ server/src/routes/patient.routes.ts
@@ -1,12 +1,9 @@
 import express from "express";
-import {
-  fetchPatients,
-  addPatient,
-  deletePatient,
-  updatePatient,
-  updatePatientUsername,
-  addMedicalTestHistory,
-} from "../controller/patient.controller";
+import fetchPatients from "../controller/patient/get.patients.controller";
+import addPatient from "../controller/patient/add.patient.controller";
+import deletePatient from "../controller/patient/delete.patient.controller";
+import updatePatient from "../controller/patient/update.patient.controller";
+import updatePatientUsername from "../controller/patient/updateName.patient.controller";
 
 const patientRouter = express.Router();
 
@@ -14,7 +11,6 @@ patientRouter.get("/", fetchPatients);
 patientRouter.post("/add", addPatient);
 patientRouter.delete("/delete", deletePatient);
 patientRouter.put("/update", updatePatient);
-patientRouter.post("/addTest", addMedicalTestHistory);
 patientRouter.patch("/patch", updatePatientUsername);
 
 export default patientRouter;
